Move diagnostics related modules into single module
This is in advance of porting Ruff's `Diagnostic` utilities over
