+++ crates/fortitude/src/check.rs
@@ -4,8 +4,7 @@ use crate::resolve;
 use crate::show_files::show_files;
 use crate::show_settings::show_settings;
 use crate::stdin::read_from_stdin;
-use fortitude_linter::diagnostic_message::DiagnosticMessage;
-use fortitude_linter::diagnostics::{Diagnostics, FixMap};
+use fortitude_linter::diagnostics::{DiagnosticMessage, Diagnostics, FixMap};
 use fortitude_linter::fs::{self, read_to_string};
 use fortitude_linter::rules::Rule;
 use fortitude_linter::rules::error::ioerror::IoError;
