+++ crates/fortitude_linter/src/diagnostics/diagnostic_message.rs
