+++ crates/fortitude_linter/src/diagnostics/message/json.rs
@@ -12,7 +12,7 @@ use ruff_diagnostics::Edit;
 use ruff_source_file::SourceCode;
 use ruff_text_size::Ranged;
 
-use crate::message::{DiagnosticMessage, Emitter};
+use super::{DiagnosticMessage, Emitter};
 
 #[derive(Default)]
 pub struct JsonEmitter;
@@ -106,8 +106,8 @@ impl Serialize for ExpandedEdits<'_> {
 mod tests {
     use insta::assert_snapshot;
 
-    use crate::message::JsonEmitter;
-    use crate::message::tests::{capture_emitter_output, create_messages};
+    use super::JsonEmitter;
+    use crate::diagnostics::message::tests::{capture_emitter_output, create_messages};
 
     #[test]
     fn output() {
