+++ crates/fortitude_server/src/lint.rs
@@ -7,7 +7,7 @@ use crate::{
     DIAGNOSTIC_NAME, PositionEncoding, edit::ToRangeExt, resolve::is_document_excluded_for_linting,
     session::DocumentQuery,
 };
-use fortitude_linter::{check_only_file, diagnostic_message::DiagnosticMessage, locator::Locator};
+use fortitude_linter::{check_only_file, diagnostics::DiagnosticMessage, locator::Locator};
 use ruff_diagnostics::{Applicability, Fix};
 use ruff_source_file::{LineIndex, SourceFileBuilder};
 use ruff_text_size::{Ranged, TextRange};
